- Removed a pair of gcc-3.0.1 compiler errors from psline.cpp
- Unlined begin() and end() functions in classes derived from touch_node.
  (I've yet to see a compiler that can inline virtual function calls on
  heap-allocated objects, anyway).  Also moved the iterator nested type
  in these classes into the source file -- it was only used by begin() and
  end().

- Made touch_node::iterator_base protected, at least for compilers that
  can cope with it being protected.

+++ examples/psline.cpp
@@ -568,9 +568,10 @@ int main(int argc, char *argv[])
     // Print the method!
     switch(args.format) {
       case eps :
-	pp = new printpage_ps(*os, 0, 0, pm.total_width(), pm.total_height() 
-			      + (args.title.empty() ? 0 
-				 : args.title_style.size * 2));
+	pp = new printpage_ps(*os, 0, 0, int(pm.total_width()), 
+			      int(pm.total_height() 
+				   + (args.title.empty() ? 0 
+				      : args.title_style.size * 2)));
 	break;
       case ps:
 	if(args.landscape) {
